requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
configparser==6.0.0
distro==1.8.0
docutils==0.20.1
gitdb==4.0.11
GitPython==3.1.40
gptchangelog @ file:///Users/xjodoin/git/gptchangelog
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
jiter==0.6.1
keyring==24.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
openai==1.52.0
pkginfo==1.9.6
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
readme-renderer==42.0
regex==2024.9.11
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
smmap==5.0.1
sniffio==1.3.0
tiktoken==0.8.0
tqdm==4.66.1
twine==5.1.0
typing_extensions==4.12.2
urllib3==2.1.0
zipp==3.17.0
